requirements.txt
pandas
openai
psycopg2-binary
python-dotenv
timescale-vector
instructor
anthropic
fastapi
uvicorn[standard]
ipykernel
pydantic
streamlit
pgvector
numpy
requests